refactored several routes
