+++ src/server/controllers/playlist.ts
@@ -20,11 +20,15 @@ export const createPlaylistURL = (req: Request, res: Response, next: NextFunctio
         })
 }
 
+// @Christina - for this function, the front end will send us the playlist name and whether they want it to be public
 // creates Spotify playlist and saves playlist ID for adding to playlist
 export const createPlaylist = (req: Request, res: Response, next: NextFunction): void => {
     const playlistURL = res.locals.playlistURL;
     const tokenData = JSON.parse(fs.readFileSync(path.join(__dirname, '../../../token.json'), 'utf8'));
     const accessToken = tokenData.access_token;
+
+    // const {name, public} = req.body;
+
     const data = {
         "name": "Test Playlist",
         "description": "A playlist generated by Emojify",
@@ -46,6 +50,7 @@ export const createPlaylist = (req: Request, res: Response, next: NextFunction):
         })
 }
 
+// @Christina - for this one we'll receive an array of tracks from the front end
 // adds tracks to playlist
 export const addTracks = (req: Request, res: Response, next: NextFunction): void => {
     const addTrackURL = res.locals.addTracksURL;
@@ -73,7 +78,7 @@ export const addTracks = (req: Request, res: Response, next: NextFunction): void
         })
 }
 
-// // no longer needed
+// // no longer needed :(
 // // fetches ID of created playlist in order to add new tracks
 // export const getPlaylistID = (req: Request, res: Response, next: NextFunction): void => {
 //     const data = JSON.parse(fs.readFileSync(path.join(__dirname, '../../../token.json'), 'utf8'));
