+++ token.json
@@ -1,1 +1,1 @@
-{"access_token":"BQBXTdbtocM_2V675xKhpE1-2BNbluFx4MlV46-I_iJxmigSPXurSk6spOZo_5PXgEuTcgUO-eMzbE1MpBC968GyaiYzUxiQgeqHZWF-Ep-_KObAfDfOm5M3QQfUgaDUtQ0PMC5ysQh8nd9rGy6PORRP_Rh3k4LlsWMP4YRLU8o_9In3Cr3HhvzZ1vM0WjEILE98PoFORYV7bFhoNLrM8zN9eF7NfPpJ9T-0QvKRrwBDKOEh6FU","token_type":"Bearer","expires_in":3600,"refresh_token":"AQAyTYN1pK0apnxJRGPCt2-05RcbgPSwFAZRq1_ANN7Mj3RQd0sApj3oZxecCRG55axrr4-OMWh5pgLHNpkXwcZcJn3ueJCwi9vOZQan_SVcwGMm8sjl9Le2PpSl851_mjg","scope":"playlist-modify-public user-read-email user-read-private"}
+{"access_token":"BQBvGQ6CgXmyuRLq2Vbkd8wx9cQoojGOzt9q0tng_eqvWVONDlgVBtYMyrc33UOgRqUYIYSfZKR9JWBX5qYCEI8kswo-k6fTQRaH2mOQNnfe1NHJVMWDA3WSX4s20A4db-p0OICUaFONqVxuyN4HasJQDLRV2raTx_tpKW47tkievK3JgRB5C9d2zJNDFpF-UyyjoB7ypcY9Cv0BiPWHg2XjAyP85Z7UbpFc8gNr93De8IH5Lbo","token_type":"Bearer","expires_in":3600,"scope":"playlist-modify-public user-read-email user-read-private","refresh_token":"AQCVIcEalunVpO0sJSrnFBuETi7NYc9FYhw8nfmJQVcyt_FiM5zN9nNSwC1wJhmxyz2yJzY39AR7SfAEU8wUll3MvOZf_3bRkQ8uw9zrU-xYhgqVqftoO6N5CfY0s-8k0zQ"}
